Refactoring execute_command -> process_speech_command in store_groceries task
