+++ mdr_planning/mdr_actions/mdr_speech_actions/mdr_process_speech_command_action/ros/launch/process_command.launch
@@ -1,4 +1,4 @@
 <?xml version="1.0"?>
 <launch>
-    <node pkg="mdr_execute_command_action" type="execute" name="process_command_server" output="screen" ns="mdr_actions" />
+    <node pkg="mdr_process_speech_command_command_action" type="process_speech_command" name="process_speech_command_server" output="screen" ns="mdr_actions" />
 </launch>
