+++ mdr_planning/mdr_scenarios/mdr_robocup_tasks/mdr_store_groceries/config/store_groceries_sm.yaml
@@ -2,7 +2,7 @@
 # Author: Alex Mitrevski
 # Email: aleksandar.mitrevski@h-brs.de
 sm_id: mdr_store_groceries
-states: [LISTEN, PROCESS_COMMAND, ENTER, GO_TO_SCANNING_POSE, FIND_TABLE, GO_TO_CUPBOARD, OPEN_DOOR, SCAN_CUPBOARD, GO_TO_TABLE, PERCEIVE_TABLE, PICK_OBJECT, FIND_OBJECTS_BEFORE_PICKING, GO_BACK_TO_CUPBOARD, PLACE_OBJECT, GO_BACK_TO_TABLE, EXIT]
+states: [LISTEN, PROCESS_SPEECH_COMMAND, ENTER, GO_TO_SCANNING_POSE, FIND_TABLE, GO_TO_CUPBOARD, OPEN_DOOR, SCAN_CUPBOARD, GO_TO_TABLE, PERCEIVE_TABLE, PICK_OBJECT, FIND_OBJECTS_BEFORE_PICKING, GO_BACK_TO_CUPBOARD, PLACE_OBJECT, GO_BACK_TO_TABLE, EXIT]
 outcomes: [DONE, TIMEOUT]
 state_descriptions:
     - state:
@@ -12,11 +12,11 @@ state_descriptions:
         transitions:
           - transition:
               name: received_command
-              state: EXECUTE_COMMAND
+              state: PROCESS_SPEECH_COMMAND
     - state:
         name: PROCESS_COMMAND
-        state_module_name: mdr_store_groceries.scenario_states.process_command
-        state_class_name: ProcessCommand
+        state_module_name: mdr_store_groceries.scenario_states.process_speech_command
+        state_class_name: ProcessSpeechCommand
         transitions:
           - transition:
               name: start_store_groceries
