+++ mdr_planning/mdr_scenarios/mdr_robocup_tasks/mdr_store_groceries/ros/launch/store_groceries.launch
@@ -17,7 +17,7 @@
 
     <include file="$(find mdr_listen_action)/ros/launch/listen.launch" />
 
-    <include file="$(find mdr_execute_command_action)/ros/launch/process_command.launch" />
+    <include file="$(find mdr_process_speech_command_action)/ros/launch/process_speech_command.launch" />
 
     <include file="$(find mdr_move_arm_action)/ros/launch/move_arm.launch" />
